refactor(concurrent): Make locking strategy granular
refactor(concurrent): Make locking strategy granular

The lock is now stored directly inside `head`, so the list and its synchronization are encapsulated together. This allows methods to use `&self`, removes the need for a separate lock field, and makes the structure more idiomatic and thread-safe.

Most importantly this allows multiple threads to access this structure
without having to be concerned with ownership, as it's our job to manage
synchronization internally.

The release + acquisition logs were removed as we will tackle doing it
in a later PR, since the structure has changed now.
- Replaced `head: Option<Box<_>>` + `lock: RwLock<()>` with `head: RwLock<Option<Box<_>>>`
- Updated mutating methods (`insert`, `delete`, `update_salary`) to take `&self` instead of `&mut self`, so they can be accessed by multiple threads
- Updated `main.rs` to remove `Mutex<HashTable>` and call methods directly on `Arc<HashTable>` - this was enforcing coarse-locking despite the intended fine-grained locking with the RwLock

Also make hashing function private, and remove derive debug from
HashTable.